+++ ChangeLog.md
@@ -3,6 +3,8 @@
  * add `--seed` argument to `smot random`
  * mix bug in proportional sampler
  * add tests of the command line utility
+ * fix bug in unfactored proportional sample
+ * allow the default factor
 
 0.1.0 [2019-11-19]